I have made the sign up and login forms work
I made sure the forms work and the input is transferred to the db
I made sure the input is sanitised and no injection can happen
I added a token to make sure that user is logged in when the form is submitted
I added validation for the form - HTML5 for frontend and PHP one for backend.
The backend consists mainly of regex - so I have set some rules for the name, username and password
I managed to fix the functions file and make sure that check_login() can check if the user is logged in and send them to the login.php if not.
I have added signup.php and login.php and they both work as expected
I redirect user to the loggedPage.php after they login and to login.php after they sign-up
I added a logout page where I unset the $_SESSION variable and redirect the user to login page.
I show a message to greet the user when they login to the loggedPage.php using their username