Completed the login page
I fixed the error messages to make them more generic
I added a function to create a random string and use it to add user ID to the db as well as token for the user and make the user be able to login only with it. It is automatically generated and saved.
I tested my form many times to try to break it and have fixed all issues that I found.
Added loggedPage.php with some variables to greet the user if successfully logged in.
I added CSS and fixed some of the existing. Made the page responsive too.
I hashed the password with password_hash and add it in a secure way to the db. I added password_verify function to my login.php page to check the hashed password and return true if the same.
Added some hints as tooltips to the signup page to inform the user what is allowed as name, username and password.
I made sure my Regex is working as expected. I added different regex for all inputs to make sure only verified data is inserted.
I added a clickable eye for the password field so the user can see what they type. Used Vanilla JS for that.